sftp_blueprints/requirements.txt
# paramiko==2.7.2
# paramiko==2.11.0
paramiko-ng==2.8.10
shipyard-utils==0.1.4
# cryptography==36.0.2
asyncssh==2.13.2